GVMScanner/requirements.txt
python-gvm
elasticsearch
